TaijiToy Color Icon Improved + Tweaks
Redid TaijiToy color icon which correctly displays the transparent option on the left side instead of giving a darker appearance on the left side. Tweaked multiple aspects on the website. The enter button now transfers to each textarea then entering on the last one submits the user input on the create order page.